requirements.txt

#python-telegram-bot
#python-telegram-bot[socks]

python-telegram-bot
requests
datetime
pyowm
matplotlib
bs4